get_futures_data/requirements.txt
airplanesdk~=0.3.14
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
